\name{eye}
\alias{eye}
\title{  eye(n) returns the n-by-n identity matrix.   eye(n,m) returns an n-by-m matrix with 1's on the diagonal and 0's elsewhere.  }
\arguments{
}
\value{List with
}
\description{
  eye(n) returns the n-by-n identity matrix.   eye(n,m) returns an n-by-m matrix with 1's on the diagonal and 0's elsewhere.  }
\author{
}
