\name{gplot}
\alias{gplot}
\title{ gplot(AdjacencyMatrix,Coordinates,Xlabel,YlabelLineSpec) as in MATLAB: plots a graph as in graph theory for an AdjacencyMatrix given the (2D or 3D )coordinates of the points }
\arguments{
\item{AdjacencyMatrix[1:n, 1:n]}{where d is2 or 3 Adjazenzmatrix des Graphen}
\item{Coordinates[1:n, 1:d]}{where d is2 or 3 coordinates of the vertices}
\item{Xlabel,Ylabel}{labels for x and y axis}
\item{LineSpec}{line colour see plot(...col =LineSpec)}
\item{Author ALU 2014}{ AdjacencyMatrix = Delaunay Coordinates = cbind(x,y) Xlabel =' ' Ylabel =' '}
}
\value{List with
}
\description{
 gplot(AdjacencyMatrix,Coordinates,Xlabel,YlabelLineSpec) as in MATLAB: plots a graph as in graph theory for an AdjacencyMatrix given the (2D or 3D )coordinates of the points }
\author{
}
