\name{normpdf}
\alias{normpdf}
\title{ computes the normal pdf at each of the values in X  R funktion fuer PDF = normpdf(X,Mean,Sdev)}
\arguments{
}
\value{List with
}
\description{
 computes the normal pdf at each of the values in X  R funktion fuer PDF = normpdf(X,Mean,Sdev)}
\author{
}
