\name{ones}
\alias{ones}
\title{ ones(n)   returns an n-by-n matrix of 1s.  ones(n,1) returns a vector of 1s  ones(n,m) returns an n-by-m matrix of ones. }
\arguments{
}
\value{List with
}
\description{
 ones(n)   returns an n-by-n matrix of 1s.  ones(n,1) returns a vector of 1s  ones(n,m) returns an n-by-m matrix of ones. }
\author{
}
