\name{polyfit}
\alias{polyfit}
\title{ finds the coefficients of a polynomial p(x) of degree n that fits the data, p(x(i)) to y(i) }
\arguments{
\item{x}{vektor mit x Werten}
\item{y}{vektor mit f(x) Werten}
\item{n}{Grad des Polynom}
\item{r}{if TRUE, use raw and not orthogonal polynomials, default: FALSE}
}
\value{p		p is a row vector of length n+1 containing the polynomial coefficients in descending powers}
\description{
 finds the coefficients of a polynomial p(x) of degree n that fits the data, p(x(i)) to y(i) }
\author{
}
