\name{polyvalALU}
\alias{polyvalALU}
\title{ function to evaluate for input vektor with input polynom output vektor of same length }
\arguments{
\item{p}{vector of polynom coefficients in descending power}
\item{x}{vector of data which should be evaluated}
}
\value{y		vector of data which is evaluated by polynom p and input data x}
\description{
 function to evaluate for input vektor with input polynom output vektor of same length }
\author{
}
