\name{RelDiff4Preise}
\alias{RelDiff4Preise}
\title{ Transformation  von Preise  in Relative Differenzen}
\arguments{
\item{prices(1:AnzTage,1:AnzKurse)}{Preise tageweise in den Spalten.}
\item{minPrice}{Preise = max(Preise,MinPreis), default MinPreis = 0.01}
}
\value{RelDiff(1:AnzTage,1:AnzKurse)}
\description{
 Transformation  von Preise  in Relative Differenzen}
\author{
}
