\name{spy}
\alias{spy}
\title{ spy(M)  plots the sparsity pattern of a  matrix M. }
\arguments{
\item{S(m,n)}{a mxn Matrix, the nonzero elements are plotted uses   PlotPixMatrix(M,Names,LowLim,HiLim)}
\item{ALU}{}
}
\value{List with
}
\description{
 spy(M)  plots the sparsity pattern of a  matrix M. }
\author{
}
