\name{waitbar}
\alias{waitbar}
\title{ close(Tacho)                                     Fortschrittsbalken schliessen  MT: Funktioniert angeblich nicht unter linux}
\arguments{
}
\value{List with
}
\description{
 close(Tacho)                                     Fortschrittsbalken schliessen  MT: Funktioniert angeblich nicht unter linux}
\author{
}
