\name{zeros}
\alias{zeros}
\title{ zeros(n)     returns an n-by-n matrix of 0s.  zeros(n,1)   returns a vector of 0s  zeros(n,m)   returns an n-by-m matrix of zeros zeros(n,m,o) returns an 3D matrix  of zeros  ALU }
\arguments{
}
\value{List with
}
\description{
 zeros(n)     returns an n-by-n matrix of 0s.  zeros(n,1)   returns a vector of 0s  zeros(n,m)   returns an n-by-m matrix of zeros zeros(n,m,o) returns an 3D matrix  of zeros  ALU }
\author{
}
